rd-45 update get-meta logic
