+++ packages/robindoc/src/types/content.ts
@@ -1,0 +1,3 @@
+export interface Provider {
+    load(uri: string): Promise<string>;
+}
